+++ NExtra/Console/ICommandLineArgumentParser.cs
@@ -4,18 +4,14 @@ namespace NExtra.Console
 {
     ///<summary>
     /// This interface can be implemented by classes that
-    /// can be used to parse command line arguments.
-    /// 
-    /// Due to the expected parse method return result, I
-    /// advice you only to use this interface when you do
-    /// expect keys and values as return values.
+    /// can parse command line args to a certain type.
     ///</summary>
     /// <remarks>
     /// Author:     Daniel Saidi [daniel.saidi@gmail.com]
     /// Link:       http://danielsaidi.github.com/nextra
     /// </remarks>
-    public interface ICommandLineArgumentParser
+    public interface ICommandLineArgumentParser<out T>
     {
-        IDictionary<string, string> ParseCommandLineArguments(IEnumerable<string> args);
+        T ParseCommandLineArguments(IEnumerable<string> args);
     }
 }