+++ NExtra/IO/IFileEncodingResolver.cs
@@ -1,19 +1,17 @@
-﻿using System.Collections.Generic;
+﻿using System.Text;
 
 namespace NExtra.IO
 {
     /// <summary>
     /// This interface can be implemented by classes that
-    /// can match file and directory paths with a pattern,
-    /// e.g. *.txt, b*, a*.txt.
+    /// can be used to resolve the text encoding of files.
     /// </summary>
     /// <remarks>
     /// Author:     Daniel Saidi [daniel.saidi@gmail.com]
     /// Link:       http://danielsaidi.github.com/nextra
     /// </remarks>
-    public interface IPathPatternMatcher
+    public interface IFileEncodingResolver
     {
-        bool IsAnyMatch(string path, IEnumerable<string> patterns);
-        bool IsMatch(string path, string pattern);
+        Encoding ResolveFileEncoding(string filePath);
     }
 }